Modify: ReviewPostArea, ReviewRow  컴포넌트 분리
