+++ src/pages/Detail/Review/ReviewPostArea/ReviewPostArea.scss
@@ -1,0 +1,1 @@
+@import '/src/styles/variables';
