+++ src/pages/Detail/Review/ReviewRow/ReviewRow.scss
@@ -1,0 +1,36 @@
+@import '/src/styles/variables';
+
+.list {
+  &:hover {
+    background-color: rgb(232, 232, 232);
+  }
+}
+
+.reviewRow {
+  display: grid;
+  grid-template-columns: 1fr 5fr 1fr 1fr;
+  padding: 20px 0 17px 0;
+  transition: 0.1s ease-in-out;
+  border-bottom: 1px solid $lightGray-color;
+  color: rgb(76, 76, 76);
+  font-size: 12px;
+
+  .reviewPadding {
+    padding: 0 30px 0 30px;
+    text-align: center;
+  }
+
+  .reviewTextboxButton {
+    padding: 0;
+    overflow: visible;
+    border: none;
+    border-radius: 0;
+    outline: none;
+    background: inherit;
+    background-color: transparent;
+    box-shadow: none;
+    font-size: 12px;
+    text-align: left;
+    cursor: pointer;
+  }
+}
